Renamed bfasster.py to run.py
